src: move test helpers into their own package
